mesh-1466: 8. extract RT commons
